Tests/LibWeb/Text/input/Internals/dump-site-isolation-process-tree.html

<!doctype html>
<script src="../include.js"></script>
<iframe srcdoc="<p>child</p>"></iframe>
<script>
    asyncTest(done => {
        addEventListener("load", () => {
            println(internals.dumpSiteIsolationProcessTree());
            done();
        });
    });
</script>
