Tests/LibWeb/Text/expected/Internals/dump-site-isolation-process-tree.txt

WebContent#0
  iframe#0: local
